Do not run database 1.3 upgrade tasks if the database didn't previously exist
git-svn-id: https://plugins.svn.wordpress.org/look-see-security-scanner/trunk@637183 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ index.php
@@ -80,7 +80,7 @@ function looksee_SQL(){
 	dbDelta($sql);
 
 	//we have some tidying up to do when upgrading to 1.0.3
-	if(get_option('looksee_db_version','0.0.0') < '1.0.3')
+	if(get_option('looksee_db_version','0.0.0') < '1.0.3' && get_option('looksee_db_version','0.0.0') !== '0.0.0')
 	{
 		//dbDelta doesn't remove old indexes, apparently
 		$wpdb->query("ALTER TABLE `{$wpdb->prefix}looksee_files` DROP INDEX `file`");
